got node space printing, brain getting tired

+++ 2023/d25-attempt/include/part2.h
@@ -1,6 +1,6 @@
 #ifndef PART2_H
 #define PART2_H
 
-unsigned int part2_solution(char *input);
+unsigned long long part2_solution(char *input);
 
 #endif
